workers/pyproject.toml
[project]
name = "workers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "groq>=0.25.0",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
    "yt-dlp>=2026.3.13",
    "boto3>=1.42.86",
]

[dependency-groups]
dev = [
    "ruff>=0.9.9",
    "pyright[nodejs]>=1.1.396",
]
